Fix multithreading issue when using Embulk with scatter-local-executor
scatter-local-executor is enabled by default when using Embulk >0.8,
which runs mulitple instance of OrcTransactionalPageOutput in parallel.

This ends up calling orc.impl.MemoryManagerImpl, which is *not*
multi-thread safe.

This change introduces WriterLocalMemoryManager, which is a simple thread-safe
implementation of orc.MemoryManager bound to each orc.Writer instance.
With this change, embulk-output-orc can be safely used from Embulk >0.8.